android try v8.4

+++ Assets/scripts/u6_slider_ctrl.cs
@@ -604,12 +604,12 @@ public class u6_slider_ctrl : MonoBehaviour
     // }
     public class SliderData
     {
-        public double Slider1;
-        public double Slider2;
-        public double Slider3;
-        public double Slider4;
-        public double Slider5;
-        public double Slider6;
+        public float Slider1;
+        public float Slider2;
+        public float Slider3;
+        public float Slider4;
+        public float Slider5;
+        public float Slider6;
     }
 
     [System.Serializable]
